+++ apps/web/app/api/analytics/[endpoint]/route.ts
@@ -5,7 +5,7 @@ import { getDomainViaEdge } from "@/lib/planetscale";
 import {
   analyticsEndpointSchema,
   getAnalyticsQuerySchema,
-} from "@/lib/zod/schemas/analytics";
+} from "@/lib/zod/schemas";
 import { NextResponse } from "next/server";
 
 // GET /api/analytics/[endpoint] – get analytics for a specific endpoint
