+++ apps/web/app/api/links/[linkId]/route.ts
@@ -3,7 +3,7 @@ import { deleteLink, editLink, processLink } from "@/lib/api/links";
 import { withAuth } from "@/lib/auth";
 import prisma from "@/lib/prisma";
 import { LinkWithTagIdsProps } from "@/lib/types";
-import { updateLinkBodySchema } from "@/lib/zod/schemas/links";
+import { updateLinkBodySchema } from "@/lib/zod/schemas";
 import { NextResponse } from "next/server";
 
 // GET /api/links/[linkId] – get a link
