+++ apps/web/app/api/links/route.ts
@@ -3,10 +3,7 @@ import { createLink, getLinksForWorkspace, processLink } from "@/lib/api/links";
 import { withAuth } from "@/lib/auth";
 import { LinkWithTagIdsProps } from "@/lib/types";
 import { ratelimit } from "@/lib/upstash";
-import {
-  createLinkBodySchema,
-  getLinksQuerySchema,
-} from "@/lib/zod/schemas/links";
+import { createLinkBodySchema, getLinksQuerySchema } from "@/lib/zod/schemas";
 import { LOCALHOST_IP } from "@dub/utils";
 import { NextResponse } from "next/server";
