+++ apps/web/app/api/tags/[id]/route.ts
@@ -2,7 +2,7 @@ import { DubApiError } from "@/lib/api/errors";
 import { withAuth } from "@/lib/auth";
 import prisma from "@/lib/prisma";
 import { recordLink } from "@/lib/tinybird";
-import { updateTagBodySchema } from "@/lib/zod/schemas/tags";
+import { updateTagBodySchema } from "@/lib/zod/schemas";
 import { NextResponse } from "next/server";
 
 // PUT /api/workspaces/[idOrSlug]/tags/[id] – update a tag for a workspace
