+++ apps/web/lib/api/links/get-links-count.ts
@@ -1,6 +1,6 @@
 import prisma from "@/lib/prisma";
 import z from "@/lib/zod";
-import { getLinksCountQuerySchema } from "@/lib/zod/schemas/links";
+import { getLinksCountQuerySchema } from "@/lib/zod/schemas";
 import { combineTagIds } from "./utils";
 
 export async function getLinksCount({
