+++ apps/web/lib/api/links/get-links-for-workspace.ts
@@ -1,6 +1,6 @@
 import prisma from "@/lib/prisma";
 import z from "@/lib/zod";
-import { getLinksQuerySchema } from "@/lib/zod/schemas/links";
+import { getLinksQuerySchema } from "@/lib/zod/schemas";
 import { combineTagIds, transformLink } from "./utils";
 
 export async function getLinksForWorkspace({
