+++ apps/web/lib/openapi/links/create-link.ts
@@ -1,5 +1,5 @@
 import { openApiErrorResponses } from "@/lib/openapi/responses";
-import { LinkSchema, createLinkBodySchema } from "@/lib/zod/schemas/links";
+import { LinkSchema, createLinkBodySchema } from "@/lib/zod/schemas";
 import { ZodOpenApiOperationObject } from "zod-openapi";
 import { workspaceParamsSchema } from "../request";
