+++ apps/web/lib/openapi/qr/get-qr.ts
@@ -1,6 +1,6 @@
 import { openApiErrorResponses } from "@/lib/openapi/responses";
 import z from "@/lib/zod";
-import { getQRCodeQuerySchema } from "@/lib/zod/schemas/qr";
+import { getQRCodeQuerySchema } from "@/lib/zod/schemas";
 import { ZodOpenApiOperationObject } from "zod-openapi";
 
 export const getQRCode: ZodOpenApiOperationObject = {
