+++ apps/web/lib/openapi/tags/get-tags.ts
@@ -1,6 +1,6 @@
 import { openApiErrorResponses } from "@/lib/openapi/responses";
 import z from "@/lib/zod";
-import { TagSchema } from "@/lib/zod/schemas/tags";
+import { TagSchema } from "@/lib/zod/schemas";
 import { ZodOpenApiOperationObject } from "zod-openapi";
 import { workspaceParamsSchema } from "../request";
