+++ apps/web/lib/zod/schemas/analytics.ts
@@ -1,7 +1,7 @@
 import { intervals, VALID_TINYBIRD_ENDPOINTS } from "@/lib/analytics";
 import z from "@/lib/zod";
 import { COUNTRY_CODES } from "@dub/utils";
-import { booleanQuerySchema } from ".";
+import { booleanQuerySchema } from "./misc";
 
 export const getAnalyticsQuerySchema = z.object({
   domain: z.string().optional().describe("The domain of the short link."),
